requirements.txt
flask
flask-cors
gunicorn
numpy
scipy
soundfile
pyyaml
werkzeug==2.2.2